stl2brundlefab.py: Refine E/Z movements
I now use relative E and Z, and at the start of the job
E is homed, and moved to feed position.

The user should then add enough powder to E to fill it, then
press the 'go' button on the front panel.

Z is left unhomed, for multi-file printing

Signed-off-by: Jason S. McMullan <jason.mcmullan@gmail.com>